inventory bulk upload

+++ backend/db/models/inventory.model.js
@@ -14,17 +14,17 @@ const Inventory = sequelize.define('inventory_prd', {
     unique: false,
   }, 
   quantity: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.FLOAT,
     allowNull: true,
     unique: false
   },
   set_quantity: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.INTEGER, 
     allowNull: true,
     unique: false
   },
   static_quantity: { // for dashboard count
-    type: DataTypes.INTEGER,
+    type: DataTypes.FLOAT,
     allowNull: true,
 
   },
